requirements-ci.txt
cupy-cuda12x>=12.0.0  # GPU tests
fakeredis>=2.20.0
faiss-cpu>=1.8.0      # fallback heavy
prometheus-client
lmdb
networkx
numpy
rich
pyyaml
fastapi
uvicorn
sqlalchemy
redis
requests
httpx
scikit-optimize
boto3
psycopg[binary]
neo4j
pybreaker
flask
flask-login
flask-wtf
wtforms
celery
opentelemetry-sdk>=1.35.0
opentelemetry-exporter-otlp>=1.35.0
opentelemetry-instrumentation-fastapi>=0.56b0
aio-pika>=9.5.5
pytest-cov
flake8
flake8-bandit
pydocstyle
docstr-coverage
radon
hypothesis>=6.98.0
jsonnet-binary
mkdocs-material>=9.5.21
bitsandbytes>=0.43.0
mypy
